app/requirements.txt
pychatgpt
ffmpeg_downloader
PyPasser
pocketsphinx
falcon
PyVirtualDisplay
selenium_profiles
python-decouple
falcon
